requirements.txt
django
djangorestframework
django-cors-headers
pyjwt
Pillow
transformers
torch
torchvision